clean up migrations, add some more testing, and api docs

+++ src/WowEvents/CombatLogVersion.cs
@@ -10,5 +10,5 @@ public class CombatLogVersion : ICombatEvent
     [LogField(1)] public string Version { get; set; } = ""; // 9
     [LogField(3)] public bool AdvancedLogEnabled { get; set; }  // 1
     [LogField(5)] public string BuildVersion { get; set; } = ""; //5.5.2
-    [LogField(7)] public int ProjectId { get; set; } // 19
+    [LogField(7)] public long ProjectId { get; set; } // 19
 }
