+++ src/WowEvents/EncounterStart.cs
@@ -7,9 +7,9 @@ public class EncounterStart : ICombatEvent
 {
     public CombatSubEventType EventType => CombatSubEventType.ENCOUNTER_START;
 
-    [LogField(1)] public int EncounterId { get; set; } // "encounterId"
+    [LogField(1)] public long EncounterId { get; set; } // "encounterId"
     [LogField(2)] public string EncounterName { get; set; } = ""; // "encounterName"
-    [LogField(3)] public int DifficultyId { get; set; } // "difficultyId"
-    [LogField(4)] public int GroupSize { get; set; } // "groupSize"
+    [LogField(3)] public long DifficultyId { get; set; } // "difficultyId"
+    [LogField(4)] public long GroupSize { get; set; } // "groupSize"
     
 }
